chore(validate country and timezone): Update validation for country in utility file to fetch from model
- update utility file in utility folder
- update validation functions for country and timezone

[Maintains #163283090]
