+++ api/block/schema.py
@@ -6,7 +6,7 @@ from api.room.schema import Room
 from api.office.models import Office
 from helpers.room_filter.room_filter import room_join_location
 from helpers.auth.admin_roles import admin_roles
-from utilities.utility import validate_empty_fields, update_entity_fields
+from utilities.validations import validate_empty_fields, update_entity_fields
 from helpers.auth.authentication import Auth
