+++ api/role/models.py
@@ -1,7 +1,8 @@
 from sqlalchemy import (Column, String, Integer)
 from sqlalchemy.schema import Sequence
 from helpers.database import Base
-from utilities.utility import Utility, validate_empty_fields
+from utilities.utility import Utility
+from utilities.validations import validate_empty_fields
 
 
 class Role(Base, Utility):
