+++ api/room_resource/schema.py
@@ -4,7 +4,7 @@ from graphql import GraphQLError
 
 from api.room_resource.models import Resource as ResourceModel
 from api.room.models import Room as RoomModel
-from utilities.utility import validate_empty_fields, update_entity_fields
+from utilities.validations import validate_empty_fields, update_entity_fields
 from helpers.auth.authentication import Auth
 from helpers.auth.admin_roles import admin_roles
 from helpers.pagination.paginate import Paginate, validate_page
