+++ api/user_role/models.py
@@ -2,7 +2,8 @@ from sqlalchemy import (Column, Integer, ForeignKey)
 from sqlalchemy.schema import Sequence
 
 from helpers.database import Base
-from utilities.utility import Utility, validate_empty_fields
+from utilities.utility import Utility
+from utilities.validations import validate_empty_fields
 
 
 class UsersRole(Base, Utility):
